+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## In progress
 
+## 1.15.3
+
 - Re-ordering for assets ([#368](https://github.com/ben/foundry-ironsworn/pull/368))
 - Re-ordering for progresses ([#369](https://github.com/ben/foundry-ironsworn/pull/369))
 - Re-ordering for connections